fix: chage tracker list was declared before loop which was carried for all the tasks
